+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 bootstrap extension Change Log
 2.0.8 under development
 -----------------------
 
+- Bug #126: Fixed `yii\bootstrap\ToggleButtonGroup` toggles dropdown for both buttons in case `split` is enabled (klimov-paul)
 - Enh #219: Add ability to use custom HTML in navbar-header (razvanphp)
 - Enh #171: Add ability to use a brandImage with the navbar (razvanphp)